plan_and_preprocess

+++ nnunetv2/imageio/simpleitk_reader_writer.py
@@ -114,14 +114,14 @@ class SimpleITKIO(BaseReaderWriter):
     def read_seg(self, seg_fname: str) -> Tuple[np.ndarray, dict]:
         return self.read_images((seg_fname, ))
 
-    def write_seg(self, seg: np.ndarray, output_fname: str, properties: dict) -> None:
+    def write_seg(self, seg: np.ndarray, output_fname: str, properties: dict, dtype=np.uint8) -> None:
         assert seg.ndim == 3, 'segmentation must be 3d. If you are exporting a 2d segmentation, please provide it as shape 1,x,y'
         output_dimension = len(properties['sitk_stuff']['spacing'])
         assert 1 < output_dimension < 4
         if output_dimension == 2:
             seg = seg[0]
 
-        itk_image = sitk.GetImageFromArray(seg.astype(np.uint8, copy=False))
+        itk_image = sitk.GetImageFromArray(seg.astype(dtype, copy=False))
         itk_image.SetSpacing(properties['sitk_stuff']['spacing'])
         itk_image.SetOrigin(properties['sitk_stuff']['origin'])
         itk_image.SetDirection(properties['sitk_stuff']['direction'])
